refactor resource API handlers
