--- internal/api/mcp_resource_read.go
@@ -1,52 +1,0 @@
-package api
-
-import (
-	"encoding/json"
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-	"github.com/mcpjungle/mcpjungle/pkg/types"
-)
-
-// getResourceHandler returns resource metadata for the given URI.
-func (s *Server) getResourceHandler() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		uri := c.Query("uri")
-		if uri == "" {
-			c.JSON(http.StatusBadRequest, gin.H{"error": "missing 'uri' query parameter"})
-			return
-		}
-
-		resource, err := s.mcpService.GetResource(uri, c.Query("server"))
-		if err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to get resource: " + err.Error()})
-			return
-		}
-
-		c.JSON(http.StatusOK, resource)
-	}
-}
-
-// readResourceHandler reads live resource content for the given URI.
-func (s *Server) readResourceHandler() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		var request types.ResourceReadRequest
-		if err := json.NewDecoder(c.Request.Body).Decode(&request); err != nil {
-			c.JSON(http.StatusBadRequest, gin.H{"error": "failed to decode request body: " + err.Error()})
-			return
-		}
-
-		if request.URI == "" {
-			c.JSON(http.StatusBadRequest, gin.H{"error": "missing 'uri' field in request body"})
-			return
-		}
-
-		resp, err := s.mcpService.ReadResource(c, request.URI, request.Server)
-		if err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to read resource: " + err.Error()})
-			return
-		}
-
-		c.JSON(http.StatusOK, resp)
-	}
-}
